ChillBro/requirements.txt
aioredis==1.3.1
amqp==2.6.1
anyjson==0.3.3
arabic-reshaper==2.1.3
asgiref==3.3.4
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
CacheControl==0.12.6
cachetools==4.2.2
celery==4.4.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
da-vinci==0.3.0
daphne==3.0.2
Django==3.2.4
django-celery-beat==2.2.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-redis==5.0.0
django-thumbnails==0.6.0
django-timezone-field==4.1.2
djangoadmin==2.2.8
djangochannelsrestframework==0.2.2
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
djangorestframework-simplejwt==4.6.0
djangotools==2.2.2
dnspython==1.16.0
eventlet==0.31.0
firebase-admin==5.0.0
freezegun==1.1.0
future==0.18.2
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.37.1
hiredis==2.0.0
html5lib==1.1
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
msgpack==1.0.2
openapi-codec==1.3.2
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.17.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==6.2.2
pytest-django==4.1.0
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.2.0
reportlab==3.5.67
requests==2.25.1
rsa==4.7.2
ruamel.yaml==0.17.0
ruamel.yaml.clib==0.2.2
service-identity==21.1.0
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
xlwt==1.3.0
zope.interface==5.4.0